SPDY assertion handler fix.
Apply a custom assertion handler to the NSURL thread that CocoaSPDY
operates on; it will print out the failure and callstack, ensure the
logs are flushed to the app, and then call abort() to send a SIGABRT
signal and crash the process.

This is necessary because an NSAssert, at least in iOS, on the NSURL
thread seems to stop its execution but not kill the process or otherwise
alert the app, as expected by an assertion failure.