Improve canonicalization.
Reduced duplication of canonicalization code by removing older version
in NSURLRequest+SPDYURLRequest.m. This means we need all requests to go
through canonicalization to ensure correctness. The request received
by the custom protocol is the canonical one in iOS 8, but is the
original request in iOS 7.

This also includes a better default User-Agent that mimics the one
produced by the NSURL system (NSURLConnection, etc).